refactor: ui와 관련된 파일을 views 폴더 하위로 수정
