+++ src/views/web/OutputView.js
@@ -5,7 +5,7 @@ import {
   getByClass,
   getById,
   showElement,
-} from '../../utils/dom.js';
+} from './utils/dom.js';
 import WinningInput from './components/WinningInput.js';
 import BonusInput from './components/BonusInput.js';
 import LottoResultModal from './components/LottoResultModal.js';
